Moved ViewModel out of StatePresenters module
