+++ src/main/java/controllers/gameStates/State.java
@@ -1,7 +1,5 @@
 package controllers.gameStates;
 
-import UI.presenters.statePresenters.ViewModel;
-
 public abstract class State {
     protected abstract void loop();
     protected abstract void render();
